// tp3/src/app/models/cv.model.ts
export interface Cv {
    [key: string]: any;

    id: number;
    name: string;
    firstname: string;
    age: number;
    cin: string;
    motto: string;
    job: string;
    description: string;
    followers: number;
    following: number;
    projects: number
    jobKeywords: string;
    path: string;
}